Extract NoneOr to separate file; Cleanup
