+++ libs/vgc/dom/schema.cpp
@@ -21,6 +21,7 @@
 #include <vgc/core/array.h>
 #include <vgc/core/colors.h>
 #include <vgc/dom/detail/facecycles.h>
+#include <vgc/dom/noneor.h>
 #include <vgc/dom/path.h>
 #include <vgc/geometry/vec2d.h>
